修复一些黄色

+++ CityO2OMerchant/Resource/Library/Calendar/CalendarLogic.m
@@ -93,7 +93,7 @@
     NSUInteger partialDaysCount = weeklyOrdinality - 1;//获取上月在这个月的日历上显示的天数
     NSDateComponents *components = [dayInThePreviousMonth YMDComponents];//获取年月日对象
     
-    for (int i = daysCount - partialDaysCount + 1; i < daysCount + 1; ++i) {
+    for (int i = (int)(daysCount - partialDaysCount + 1); i < daysCount + 1; ++i) {
         
         CalendarDayModel *calendarDay = [CalendarDayModel calendarDayWithYear:components.year month:components.month day:i];
         calendarDay.style = CellDayTypeEmpty;//不显示
@@ -278,7 +278,7 @@
 -(void)LunarForSolarYear:(CalendarDayModel *)calendarDay{
     
     
-    NSString *solarYear = [self LunarForSolarYear:calendarDay.year Month:calendarDay.month Day:calendarDay.day];
+    NSString *solarYear = [self LunarForSolarYear:(int)calendarDay.year Month:(int)calendarDay.month Day:(int)calendarDay.day];
     
     NSArray *solarYear_arr= [solarYear componentsSeparatedByString:@"-"];
